refactor: remove legacy hardware diagnostics path
